evaluate error condition codes
